indenpendent security session factory

+++ security/org.beangle.security.core/src/main/java/org/beangle/security/core/session/SessionController.java
@@ -63,6 +63,6 @@ public interface SessionController {
    * @param auth
    * @return None or inactive interval
    */
-  Option<Integer> getInactiveInterval(Authentication auth);
+  Option<Short> getInactiveInterval(Authentication auth);
 
 }
