+++ security/org.beangle.security.core/src/main/java/org/beangle/security/core/session/category/CategoryProfile.java
@@ -44,11 +44,11 @@ public interface CategoryProfile {
   /**
    * Max session for same account
    */
-  int getUserMaxSessions();
+  short getUserMaxSessions();
 
   /**
    * Max inactiveInterval in minutes
    */
-  int getInactiveInterval();
+  short getInactiveInterval();
 
 }
